backend/forecasting_service/requirements.txt
Flask
pandas
prophet
pymongo
python-dotenv
gunicorn